WIP: winebus: Guess the type of evdev input devices
Ordinarily, we can get the type of an evdev input device from udev:
the input_id builtin sets udev properties of the form ID_INPUT_FOO
that we can read.

However, in a container there is no guarantee that the libudev in the
container will interoperate with the udevd on the host system, so we
need to be prepared to do this ourselves from first principles, using
a heuristic similar to the one in udev's input_id.

We cannot simply copy the heuristic from udev's input_id, because its
licensing is incompatible (GPL). Instead, use a vaguely similar heuristic
that works from the same inputs and will generally produce similar results.

TODO: This code was taken from the steam-runtime-tools library. To make it
easier to keep the heuristic in sync with that library, it has only been
minimally changed for inclusion here. When we are happy with the heuristic,
we should reformat this code to fit Wine coding style, removing or inlining
the GLib compatibility glue.

Signed-off-by: Simon McVittie <smcv@collabora.com>